fix(click-effects): Improve click detection and positioning
Updates `ClickEffects` to use `clientX/Y` for accurate positioning relative to the viewport and adds `{ capture: true }` to the event listener. This ensures click effects are consistently rendered and triggered, even on elements that stop event propagation, and makes the effect container fixed to the viewport.

Also updates the `index.html` to uncomment the importmap and improve the polyfill comment for `process.env`.